WIP: Break out extract dataset info util
